+++ sample/proto_v2/role_cfg.n.js
@@ -3,7 +3,7 @@
 exports.role_cfg_header = {
   xres_ver : "1.4.3",
   hash_code : "no hash code",
-  data_ver : "1.4.3.20180329102023",
+  data_ver : "1.4.3.20180330124812",
   count : 7
 };
 exports.role_cfg = [